Ignore P4DIFF and DIFF in p4-revert
p4-revert runs 'p4 diff' and expects a textual diff response. This
assumption is violated if P4DIFF points to a graphical diff tool, which
may not return a textual diff. The same behavior happens if P4DIFF is not
set and DIFF points to a graphical diff tool, since p4 falls back to using
DIFF.

We fix this by clearing the P4DIFF and DIFF variables from the local
environment before running 'p4 diff'. This behavior change only applies
to the p4-revert function.