Escape @ characters as %40
We use Matlab, Matlab loves using @ in direcory names, Perforce does
not. This will simple replace the @ with the %40 escape code.

If i was better at elisp i would handle every character which needs to
be escaped:

https://www.perforce.com/perforce/r12.1/manuals/cmdref/o.fspecs.html

I would probably also add a p4-escape-name function which does the
work.